+++ src/usbmon/mod.rs
@@ -7,6 +7,8 @@ use std::process::Command;
 use std::time::Duration;
 
 pub mod binary;
+#[cfg(feature = "ebpf")]
+mod ebpf;
 pub mod mmap_ring;
 pub mod monitor;
 pub mod parser;